ibmvnic: Fix overflowing firmware/hardware TX queue
Use a counter torack the number of outstanding transmissions sent
that have not received completions. If we get to close to the number of
queue entries agreed upon during resource negotiation, stop
transmissions on that queue. As we receive more completions from
firmware, wake the queue once the counter reaches an acceptable level.