ibmvnic: Update failover handling to use close/open
The handling of a failover condition in the driver needs to release
the resources for the driver and re-acquire them. Update the
failover handling code to use the updated close and open routiunes to
do this.

Additionally this patch re-names the workqueue for failover handling
as failover is the only action we use the workqueue for.

Signed-off-by: Nathan Fontenot <nfont@linux.vnet.ibm.com>