ibmvnic: Reset adapter for fatal errors
Tom,

I've moved the reset into the existing workqueue. Not sure if we need
the two second delay on this or not, but kept it for consistency
with the other places where we schedule work on this.

8<

This fixes up ibmvnic to properly reset the vnic interface
if a fatal error is received as is defined as required
by the PAPR+ interface.

Signed-off-by: Brian King <brking@linux.vnet.ibm.com>