ibmvnic: Correct resource handling
The resources allocated by the vnic driver are mis-handled in several places.
There are instances of when we release resources but do not need to, and times
when resources need to be reset but are not.

This patch updates the open, close, and remove routines for the driver to do
the proper handling of resources.

Signed-off-by: Nathan Fontenot <nfont@linux.vnet.ibm.com>