ibmvnic: Move init and login out of __ibmvnic_open
Move the initialization of vnic resources and login to the vios server
out of the __ibmvnic_open routine and make them a separate step as
part of the boot up process for the driver. This gives us a finer
grained control on where we are with adapter status when handling
resets.

Signed-off-by: Nathan Fontenot <nfont@linux.vnet.ibm.com>