ibmvnic: handle getting a link state cloised on link state send
When setting the link state during a reset it is possible that the
vnicserver we fail over to is not available. In this case the send_crq
call will return H_CLOSED. Update the set_link_state to check the
send_crq return code and handle H_CLOSED by calling ibmvnic_remove.