ibmvnic: Correct tx pool cleaning during close
If the NetworkManager is not started on a system and a vnic device
is available, the vnic driver will have the probe routine called
but not the open routine. This is expected behavior.

This will cause an issue if we get a failover. During the failover
handling we call __ibmvnic_close() which attempts to clear any
hanging skbs. Since the driver was never opened, the tx pools were
never allocated, and there is nothing to clean up.

The current code only checks adapter->tx_scrqs before doing this,
but the sub crqs are initialized during probe so this is not the
correct protection for cleaning the tx pools.

This patch corrects this to check for the existence of tx pools
before trying to clean any skbs. The patch also moves the disabling
of the irqs for the tx and rx sub-crqs back to disable_sub_crqs().

Signed-off-by: Nathan Fontenot <nfont@linux.vnet.ibm.com>