ibmvnic: Correctly handle passive inits
Latest version of the passive init patch. This is pretty close to the way I'd
like it to be in the long run. To me it seems ideal to handle passive inits
consistently. This version of the patch almost achieves this. In each passive
init scenario, we simply queue a failover reset task. Based on my testing,
this handles each scenario correctly.

This patch also refines the hack to handle a failover in the middle of an eeh.

Signed-off-by: John Allen <jallen@linux.vnet.ibm.com>